+++ babel.config.js
@@ -1,51 +1,7 @@
 module.exports = api => {
   api.cache(true);
-  const plugins = [
-    '@babel/plugin-syntax-dynamic-import',
-    '@babel/plugin-syntax-import-meta',
-    '@babel/plugin-proposal-class-properties',
-    [
-      '@babel/plugin-proposal-decorators',
-      {
-        legacy: true,
-      },
-    ],
-    '@babel/plugin-proposal-function-sent',
-    '@babel/plugin-proposal-export-namespace-from',
-    '@babel/plugin-proposal-numeric-separator',
-    '@babel/plugin-proposal-throw-expressions',
-    '@babel/plugin-proposal-export-default-from',
-    '@babel/plugin-proposal-logical-assignment-operators',
-    '@babel/plugin-proposal-optional-chaining',
-    [
-      '@babel/plugin-proposal-pipeline-operator',
-      {
-        proposal: 'minimal',
-      },
-    ],
-    '@babel/plugin-proposal-nullish-coalescing-operator',
-    '@babel/plugin-proposal-do-expressions',
-    '@babel/plugin-proposal-function-bind',
-    'babel-plugin-jest-hoist',
-  ];
-  const presets = [
-    '@babel/preset-react',
-    [
-      '@babel/preset-env',
-      {
-        debug: false,
-        modules: 'commonjs',
-        useBuiltIns: 'usage',
-        corejs: 3,
-        targets: {
-          node: 'current',
-        }
-      },
-    ],
-  ];
 
   return {
-    presets,
-    plugins,
+    presets: ['babel-preset-react-app'],
   };
 };