+++ packages/date/tests/DateRange.test.js
@@ -281,4 +281,43 @@ describe('DateRange', () => {
       );
     });
   });
+
+  test('pre-selected date ranges appear', async () => {
+    const onSubmit = jest.fn();
+
+    const { container } = render(
+      <Form
+        initialValues={{
+          dateRange: undefined,
+        }}
+        onSubmit={onSubmit}
+      >
+        <DateRange id="dateRange" name="dateRange" ranges />
+        <Button type="submit">Submit</Button>
+      </Form>
+    );
+
+    container.querySelector('.DateRangePickerInput_calendarIcon').click();
+
+    await wait(() => {
+      expect(
+        container.querySelector(
+          '.DayPicker_calendarInfo__horizontal DayPicker_calendarInfo__horizontal_1'
+        )
+      ).toBeDefined();
+    });
+
+    // Simulate User hitting the 'Today' pre-set
+    container.querySelectorAll('.btn-default')[0].click();
+
+    const date = new Date();
+
+    expect(
+      container.querySelectorAll('.DateInput_input.DateInput_input_1')[0].value
+    ).toEqual(`${date.getMonth() + 1}/${date.getDate()}/${date.getFullYear()}`);
+
+    expect(
+      container.querySelectorAll('.DateInput_input.DateInput_input_1')[1].value
+    ).toEqual(`${date.getMonth() + 1}/${date.getDate()}/${date.getFullYear()}`);
+  });
 });