+++ packages/favorites/types/FavoriteHeart.d.ts
@@ -1,0 +1,8 @@
+import { MouseEventHandler } from 'react-select/src/types';
+
+interface FavoriteHeartProps extends React.HTMLAttributes<HTMLSpanElement> {
+  id: string;
+}
+declare const FavoriteHeart: React.FunctionComponent<FavoriteHeartProps>;
+
+export default FavoriteHeart;