+++ packages/reactstrap-validation-select/AvResourceSelect.js
@@ -13,40 +13,6 @@ class AvResourceSelect extends Component {
     this.state.previousOptions = [];
   }
 
-  static propTypes = {
-    requestConfig: PropTypes.object,
-    resource: PropTypes.shape({
-      postGet: PropTypes.func,
-      post: PropTypes.func,
-      getResult: PropTypes.oneOfType([PropTypes.string, PropTypes.func]),
-    }).isRequired,
-    getResult: PropTypes.oneOfType([PropTypes.string, PropTypes.func]),
-    hasMore: PropTypes.oneOfType([PropTypes.bool, PropTypes.func]),
-    delay: PropTypes.number,
-    debounceTimeout: PropTypes.number,
-    label: PropTypes.node,
-    customerId: PropTypes.string,
-    parameters: PropTypes.object,
-    itemsPerPage: PropTypes.number,
-    onPageChange: PropTypes.func,
-    isDisabled: PropTypes.bool,
-    requiredParams: PropTypes.array,
-    watchParams: PropTypes.array,
-    // eslint-disable-next-line react/forbid-prop-types
-    cacheUniq: PropTypes.any,
-    additional: PropTypes.object,
-    graphqlConfig: PropTypes.shape({
-      type: PropTypes.string,
-      query: PropTypes.string,
-    }),
-    minCharsToSearch: PropTypes.number,
-  };
-
-  static defaultProps = {
-    delay: 350,
-    itemsPerPage: 50,
-  };
-
   select = createRef();
 
   loadOptions = (...args) => {
@@ -247,6 +213,40 @@ AvResourceSelect.create = defaults => {
     defaults.resource.defaultConfig.name
   )}Select`;
   return SpecificAvResourceSelect;
+};
+
+AvResourceSelect.propTypes = {
+  requestConfig: PropTypes.object,
+  resource: PropTypes.shape({
+    postGet: PropTypes.func,
+    post: PropTypes.func,
+    getResult: PropTypes.oneOfType([PropTypes.string, PropTypes.func]),
+  }).isRequired,
+  getResult: PropTypes.oneOfType([PropTypes.string, PropTypes.func]),
+  hasMore: PropTypes.oneOfType([PropTypes.bool, PropTypes.func]),
+  delay: PropTypes.number,
+  debounceTimeout: PropTypes.number,
+  label: PropTypes.node,
+  customerId: PropTypes.string,
+  parameters: PropTypes.object,
+  itemsPerPage: PropTypes.number,
+  onPageChange: PropTypes.func,
+  isDisabled: PropTypes.bool,
+  requiredParams: PropTypes.array,
+  watchParams: PropTypes.array,
+  // eslint-disable-next-line react/forbid-prop-types
+  cacheUniq: PropTypes.any,
+  additional: PropTypes.object,
+  graphqlConfig: PropTypes.shape({
+    type: PropTypes.string,
+    query: PropTypes.string,
+  }),
+  minCharsToSearch: PropTypes.number,
+};
+
+AvResourceSelect.defaultProps = {
+  delay: 350,
+  itemsPerPage: 50,
 };
 
 export default AvResourceSelect;