+++ packages/reactstrap-validation-select/AvSelect.js
@@ -6,6 +6,7 @@ import Select, { components as reactSelectComponents } from 'react-select';
 import Creatable from 'react-select/creatable';
 import get from 'lodash.get';
 import isEqual from 'lodash.isequal';
+import isFunction from 'lodash.isfunction';
 
 import AsyncPaginate from 'react-select-async-paginate';
 
@@ -44,18 +45,6 @@ class AvSelect extends AvBaseInput {
       newOptions: [],
     };
   }
-
-  static contextTypes = {
-    FormCtrl: PropTypes.object,
-  };
-
-  static propTypes = Object.assign({}, AvBaseInput.propTypes, {
-    options: PropTypes.array,
-    loadOptions: PropTypes.func,
-    raw: PropTypes.bool,
-    creatable: PropTypes.bool,
-    autofill: PropTypes.oneOfType([PropTypes.bool, PropTypes.object]),
-  });
 
   optionsContainsValue = props => {
     const valueKey = this.getValueKey(props);
@@ -213,11 +202,17 @@ class AvSelect extends AvBaseInput {
 
             let val;
             if (typeof this.props.autofill === 'object') {
-              val = get(
-                rawValue,
-                `${this.props.autofill[fieldName]}`,
-                input.getDefaultValue()
-              );
+              if (isFunction(this.props.autofill[fieldName])) {
+                val = this.props.autofill[fieldName](rawValue);
+              } else if (typeof this.props.autofill[fieldName] === 'string') {
+                val = get(
+                  rawValue,
+                  `${this.props.autofill[fieldName]}`,
+                  input.getDefaultValue()
+                );
+              } else {
+                val = input.getDefaultValue();
+              }
             } else {
               val = get(rawValue, fieldName, input.getDefaultValue());
             }
@@ -322,9 +317,20 @@ class AvSelect extends AvBaseInput {
             ...provided,
             width: '90%',
           }),
+          singleValue: provided => {
+            return {
+              ...provided,
+              color: '#495057',
+            };
+          },
           control: (provided, state) => {
             if (state.isDisabled) {
-              return provided;
+              return {
+                ...provided,
+                borderRadius: '.25em',
+                borderColor: '#ced4da',
+                backgroundColor: '#e9ecef',
+              };
             }
             const showError = touched && hasError && !state.focused;
 
@@ -369,5 +375,18 @@ class AvSelect extends AvBaseInput {
     );
   }
 }
+
+AvSelect.contextTypes = {
+  FormCtrl: PropTypes.object,
+};
+
+AvSelect.propTypes = {
+  ...AvBaseInput.propTypes,
+  options: PropTypes.array,
+  loadOptions: PropTypes.func,
+  raw: PropTypes.bool,
+  creatable: PropTypes.bool,
+  autofill: PropTypes.oneOfType([PropTypes.bool, PropTypes.object]),
+};
 
 export default AvSelect;