+++ packages/mock/CHANGELOG.md
@@ -3,6 +3,17 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.0.2](https://github.com/Availity/availity-react/compare/@availity/mock@1.0.1...@availity/mock@1.0.2) (2019-10-18)
+
+
+### Bug Fixes
+
+* **mock:** fixed incorrect slotmachine url ([9050f3d](https://github.com/Availity/availity-react/commit/9050f3d))
+
+
+
+
+
 ## 1.0.1 (2019-09-25)
 
 **Note:** Version bump only for package @availity/mock