+++ packages/upload/CHANGELOG.md
@@ -3,6 +3,17 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [2.1.3](https://github.com/Availity/availity-react/compare/@availity/upload@2.1.2...@availity/upload@2.1.3) (2019-10-11)
+
+
+### Bug Fixes
+
+* eslint issues ([bc4d8e9](https://github.com/Availity/availity-react/commit/bc4d8e9))
+
+
+
+
+
 ## [2.1.2](https://github.com/Availity/availity-react/compare/@availity/upload@2.1.1...@availity/upload@2.1.2) (2019-10-01)
 
 **Note:** Version bump only for package @availity/upload